+++ src/prepress/__init__.py
@@ -3,4 +3,4 @@ from importlib.metadata import version, PackageNotFoundError
 try:
     __version__ = version("prepress")
 except PackageNotFoundError:
-    __version__ = "0.1.2"
+    __version__ = "0.2.0"